fixes on new compression algo, cleanup
